[refactor] Rearrange util files to usecases
