--- src/usecases/scoreUtils.ts
@@ -1,3 +1,0 @@
-export const getHighScore = (newScore: number, prevHighScore: number) => {
-  return newScore > prevHighScore ? newScore : prevHighScore;
-};
